Split cluster.Config
...into multiple files, right now:
- "main" holding the iverall structure
- "kubeadm" holding all the configs from the kubeadm config struct
- "lightweight" holding the additional configs for the lightweight
  framewaork
